fix(rust-analyzer): have rust-tools setup rust-analyzer
Loading "rust-tools" automatically sets up rust-analyzer.  However, to
modify rust-analyzer's config, this has to happen *when* rust-tools is
loaded.

As part of the fix, move Lua LSP setup into its own function that does
not rely on the default handler.  This is required as `lua_ls` requires
a "settings" key for configurations, which is not true of other
handlers.